Refactor location of tests
