--- __tests__/components/App.test.js
@@ -1,12 +1,0 @@
-import React from 'react';
-
-
-
-
-describe('App.test.js', () => {
-    it('renders without crashing', () => {
-
-    });
-
-
-});
