+++ src/components/common/__tests__/FieldInput.test.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { shallow } from 'enzyme';
 import toJson from 'enzyme-to-json';
-import FieldInput from '../../../src/components/common/FieldInput';
+import FieldInput from '../FieldInput';
